WIP Assistants Streaming: Updated types
- Updated types
- Removed unnecessary error in StreamedThreadRunResponseFactory
- Ensured the types match the API responses correctly
