+++ src/Responses/Threads/Messages/Delta/ThreadMessageDeltaResponseContentImageFileObject.php
@@ -7,14 +7,15 @@ namespace OpenAI\Responses\Threads\Messages\Delta;
 use OpenAI\Contracts\ResponseContract;
 use OpenAI\Responses\Concerns\ArrayAccessible;
 use OpenAI\Testing\Responses\Concerns\Fakeable;
+use OpenAI\Testing\Responses\Threads\Messages\ThreadMessageResponseContentImageFile;
 
 /**
- * @implements ResponseContract<array{type: string, image_file: array{file_id: string}}>
+ * @implements ResponseContract<array{index: int, type: string, image_file: array{file_id: string}}>
  */
 final class ThreadMessageDeltaResponseContentImageFileObject implements ResponseContract
 {
     /**
-     * @use ArrayAccessible<array{type: string, image_file: array{file_id: string}}>
+     * @use ArrayAccessible<array{index: int, type: string, image_file: array{file_id: string}}>
      */
     use ArrayAccessible;
 
@@ -23,21 +24,21 @@ final class ThreadMessageDeltaResponseContentImageFileObject implements Response
     private function __construct(
         public ?int $index,
         public string $type,
-        public ThreadMessageDeltaResponseContentImageFile $imageFile,
+        public ThreadMessageResponseContentImageFile $imageFile,
     ) {
     }
 
     /**
      * Acts as static factory, and returns a new Response instance.
      *
-     * @param  array{type: string, image_file: array{file_id: string}}  $attributes
+     * @param  array{index: int, type: string, image_file: array{file_id: string}}  $attributes
      */
     public static function from(array $attributes): self
     {
         return new self(
             isset($attributes['index']) ? $attributes['index'] : null,
             $attributes['type'],
-            ThreadMessageDeltaResponseContentImageFile::from($attributes['image_file']),
+            ThreadMessageResponseContentImageFile::from($attributes['image_file']),
         );
     }
